fix(markdown): move withShiki to its own entry point
shiki was imported from the markdown FESM, and provideNgeDoc's default
renderer loads that FESM to render pages, so every nge-doc consumer had to
install shiki even when it never called withShiki (the build failed with
"Could not resolve shiki").

Move withShiki and the shiki service to @cisstech/nge/markdown/shiki, so
shiki is pulled only by apps that opt into it. The default renderer no longer
forces it.

Breaking: import withShiki from '@cisstech/nge/markdown/shiki' instead of
'@cisstech/nge/markdown'.
