+++ projects/nge/markdown/index.ts
@@ -5,7 +5,6 @@ export * from './src/contributions/nge-markdown-highlighter'
 export * from './src/contributions/nge-markdown-icons'
 export * from './src/contributions/nge-markdown-katex'
 export * from './src/contributions/nge-markdown-link-anchor'
-export * from './src/contributions/nge-markdown-shiki'
 export * from './src/contributions/nge-markdown-tabbed-set'
 export * from './src/nge-markdown-transformer'
 export * from './src/nge-markdown-components'
