+++ projects/nge/markdown/shiki/index.ts
@@ -1,0 +1,1 @@
+export * from './shiki'
